Remove venv config file after failure update.sh script
Sometimes the venv creation does exceed, but the upgrade.sh script does
fail after the venv is created. By removing the venv config file, which
is used in the unless command of the exec, the exec will be run again on
the next puppet run.

This can happen during postgresql or python upgrades triggered by this
module.